+++ SaturnEngine/SEUIControls/SEPanel.cs
@@ -16,7 +16,7 @@ namespace SaturnEngine.SEUIControls
             
         }
 
-        public override void Update(float deltaTime)
+        public override void Update(double deltaTime)
         {
 
         }
